requirements.txt
PyQt5==6.5.2
openai==1.1.0
requests==2.31.0
python-dotenv==1.0.1
sqlite3==2.6.0
markdown==3.5.1
pygments==2.16.1


